Added some of the Basic Files + Experimentation
Here I have added some of the basic files as well as some experimentation for the overall game. I used Ursina's built in First Person Controller to create the base game. So far I have found a way to remove V-Sync, change the speed of the character and the size of the map.  I am yet to add any documentation for this, however this will be accounted for today.